task 13/requirements.txt
flask
pandas
sentence-transformers
faiss-cpu
numpy
